Update ArduinoDRO.ino
Added visual feedback to operator about lengths of belts in measurement vs history of belts we have measured. The bench calls a belt good if it is within -0.007" / +0.010" of our current "control" ExactFit belts for each model. 

Do note that the global beltMax[] array has the maximum observed c2c values of used OEM belts (when available). The OEM 900 belt is allowed 0.020" longer c2c than we call good and the OEM996 belt has been observed at an astounding 0.040" longer center to center. :P

Updated the code in calibrate() in the upload section to remind you to unplug the wall wart for uploading. This way we don't have to unseat the 7166 IC, which is a PITA.